+++ lib/src/01_logging.dart
@@ -1,11 +1,11 @@
-// Copyright (c) 2021 Artёm Galkin. All rights reserved.
+// Copyright (c) 2021 Artёm Galkin <github.com/rtmigo>. All rights reserved.
 // Use of this source code is governed by the BSD-3-Clause license found
 // in the LICENSE file in the root directory of this source tree
 
 import '00_global.dart';
 
 void logPrint(String txt) {
-  print("prefnotifiers: $txt");
+  print('prefnotifiers: $txt');
 }
 
 void logInfo(String txt) {