+++ lib/src/10_awaitableCalls.dart
@@ -1,10 +1,11 @@
-// Copyright (c) 2021 Artёm Galkin. All rights reserved.
+// Copyright (c) 2021 Artёm Galkin <github.com/rtmigo>. All rights reserved.
 // Use of this source code is governed by the BSD-3-Clause license found
 // in the LICENSE file in the root directory of this source tree
 
 import 'dart:async';
 
-/// Controls the asynchronous read and write process so that reads occur only after all writes have completed.
+/// Controls the asynchronous read and write process so that reads occur
+/// only after all writes have completed.
 ///
 /// To create object:
 /// ```
@@ -22,10 +23,10 @@ import 'dart:async';
 ///   await readFunc();
 /// ```
 class AwaitableCalls {
-  List<Completer> _items = List<Completer>();
+  final List<Completer> _items = <Completer>[];
 
   Future run(func) async {
-    Completer c = Completer();
+    Completer? c = Completer();
     this._items.add(c);
 
     try {
@@ -35,12 +36,16 @@ class AwaitableCalls {
       c = null;
       rethrow;
     } finally {
-      if (c != null) c.complete();
+      if (c != null) {
+        c.complete();
+      }
       this._items.remove(c);
     }
   }
 
   Future<void> completed() async {
-    while (this._items.length > 0) await this._items[0].future;
+    while (this._items.isNotEmpty) {
+      await this._items[0].future;
+    }
   }
 }