+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.0.8+1
+
+- FIX: An attempt to save or read a `List<String>` raised an exception
+
+# 1.0.7+1
+
+- Not a beta anymore
+- Documentation updates 
+
+# 1.0.2-beta
+
+- Migrated to sound null safety
+- Added `PrefNotifier` class as a shortcut to `PrefItem(SharedPrefsStorage(), ...)`
+- Multiple minor improvements 
+
 # 0.0.7
 
 - Example simplified